Move CopyButton to base components
