+++ src/components/base/button/index.ts
@@ -1,0 +1,4 @@
+export { default as CopyButton } from './CopyButton.svelte';
+
+import Button from './Button.svelte';
+export default Button;
