+++ src/components/base/index.ts
@@ -1,4 +1,4 @@
-export { default as Button } from './Button.svelte';
+export { default as Button } from './button';
 export { default as Select } from './Select.svelte';
 export { default as TextInput } from './TextInput.svelte';
 export { default as TextArea } from './TextArea.svelte';
